+++ threadx_test/evaluation/jsonCollect.py
@@ -1,4 +1,5 @@
 import os
+import re
 directory = "/home/klee/threadx/symbolic_execution/test-info-output"
 
 def list_json_files(directory):
@@ -11,13 +12,24 @@ def list_json_files(directory):
 
     return json_files
 
+def extract_api_info(file_path):
+    pattern = r"/description_(.+?)_(\d+)_(\w+)\.json$"
+    match = re.search(pattern, file_path)
+    if match:
+        return match.group(1), match.group(3)
+    else:
+        return None, None
+
 def main():
     
     json_files = list_json_files(directory)
 
     with open("jsonFilesPath.txt", "w") as f:
         for file_path in json_files:
+            api_name, mo_name = extract_api_info(file_path)
             f.write(file_path + '\n')
+            f.write(api_name + '\n')
+            f.write(mo_name + '\n')
 
     print(f"All .json files have been recorded to jsonFilesPath.txt.")