+++ threadx_test/evaluation/run_test.py
@@ -78,9 +78,10 @@ def run_command(func_name):
 
 print(func_names)
 
-P = Pool(processes=1)
-P.map(run_command,func_names)
-
+# P = Pool(processes=1)
+# P.map(run_command,func_names)
+for func_name in func_names:
+    run_command(func_name)
 
 # with concurrent.futures.ProcessPoolExecutor(max_workers=2) as executor:
 #     executor.map(run_command, func_names)